homework/Class 3 HomeWork/voice-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
torch==2.1.0
transformers==4.35.0
pydub==0.25.1
numpy==1.24.3
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0
gtts==2.4.0
python-dotenv==1.0.0



